requirements.txt
streamlit
pyaudio
wave
openai
google.generativeai
python-dotenv
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.3
asttokens==2.4.1
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.36.13
botocore==1.36.13
cachetools==5.5.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
executing==2.0.1
fastapi==0.115.8
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.7
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.7.0
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mdurl==0.1.2
narwhals==1.24.0
nest-asyncio==1.5.8
numpy==2.0.0
openai==1.60.1
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==5.29.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
pyzmq==25.1.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.45.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
Wave==0.0.2
wcwidth==0.2.12
langchain
langchain-openai
langchain-google-genai
langchain_experimental
openai
anthropic
cohere
streamlit
python-dotenv
beautifulsoup4
faiss-cpu
pypdf
unstructured
networkx
openpyxl
rapidocr-onnxruntime
streamlit
pyaudio
wave
openai
langchain
langchain-openai
langchain-google-genai
langchain_experimental
faiss-cpu
fastapi
uvicorn
whisper
python-dotenv
annotated-types
numpy
pandas
altair
beautifulsoup4
requests
tenacity
jupyter_core
protobuf
starlette
tqdm
google-auth
google-api-python-client
google-generativeai
openai-whisper
torch
torchaudio

